Fix imports and refactor packages
